Chemist v0.1.0 Chemist.Game

Retrieve game data from Riot API and transform it into an Elixir friendly format

Functions

current(region, player_id)

Return a tuple with the values { :ok, data } where data is a map containing current game data.

Uses “current-game-v1.0” API.

featured(region)

Return a tuple with the values { :ok, data } where data is a map containing featured game data.

Uses “featured-games-v1.0” API.

recent(region, player_id)

Return a tuple with the values { :ok, data } where data is a list of maps containing the summoner’s recent game data.

Uses “game-v1.3” API.