Chemist v0.1.0 Chemist.League

Uses league-v2.5 API.

Functions

challenger(region, opts \\ %{})

Contains league entries for the challenger league.

Default opts:

type: “RANKED_FLEX_SR”

Type of matchmaking

Possible values:

RANKED_FLEX_TT

RANKED_SOLO_5x5

RANKED_TEAM_3x3

RANKED_TEAM_5x5

Sample output:

{:ok,
 %{"entries" => [%{"division" => "I", "isFreshBlood" => true,
      "isHotStreak" => true, "isInactive" => false, "isVeteran" => false, 
      "leaguePoints" => 0, "losses" => 18,
      "playerOrTeamId" => "25270227", "playerOrTeamName" => "DusKRaptor", 
      "wins" => 43},
      ...
      "name" => "Poppy's Paladins", "queue" => "RANKED_FLEX_SR",
      "tier" => "CHALLENGER"}}

entry_by_id(region, player_id)

Contains all league entries for specified summoners and summoners’ teams; retrieved by .

Sample output:

{:ok,
 %{"51666047" => [%{"entries" => [%{"division" => "V",
         "isFreshBlood" => true, "isHotStreak" => false,
         "isInactive" => false, "isVeteran" => false,
         "leaguePoints" => 21, "losses" => 133,
         "playerOrTeamId" => "51666047",
         "playerOrTeamName" => "jrizznezz", "wins" => 139}],
      "name" => "Draven's Renegades", "queue" => "RANKED_SOLO_5x5",
      "tier" => "GOLD"}]}}

league_by_id(region, player_id)

Contains all leagues for specified summoners and summoners’ teams, including for teams where the player is inactive; retrieved by player id.

Sample output:

{:ok,
 %{"51666047" => [%{"entries" => [%{"division" => "IV",
         "isFreshBlood" => false, "isHotStreak" => false,
         "isInactive" => false, "isVeteran" => false,
         "leaguePoints" => 0, "losses" => 7,
         "playerOrTeamId" => "24883051",
         "playerOrTeamName" => "The Bootyologist", "wins" => 4},
         ...
         "name" => "Draven's Renegades", "participantId" => "51666047",
              "queue" => "RANKED_SOLO_5x5", "tier" => "GOLD"}]}}

master(region, opts \\ %{})

Contains league entries for the master league.

Default opts:

type: “RANKED_FLEX_SR”

Type of matchmaking

Possible values:

RANKED_FLEX_TT

RANKED_SOLO_5x5

RANKED_TEAM_3x3

RANKED_TEAM_5x5

Sample output:

{:ok,
 %{"entries" => [%{"division" => "I", "isFreshBlood" => true,
      "isHotStreak" => false, "isInactive" => false,
      "isVeteran" => false, "leaguePoints" => 0, "losses" => 36,
      "playerOrTeamId" => "38809624",
      "playerOrTeamName" => "Sophist Sage", "wins" => 49}],
      ...
   "name" => "Darius's Blackguards", "queue" => "RANKED_FLEX_SR",
   "tier" => "MASTER"}}