Chemist v0.1.0 Chemist.Static

Retrieve league data from Riot API and transform it into an Elixir friendly format

Functions

champion(region, champion_id, opts \\ %{})

Contains champion data; retrieved by champion id.

Default opts:

locale: default varies based on region

Locale code for returned data (e.g., en_US, es_ES).

version: latest version based on region

Data dragon version for returned data.

champData: Only type, version, data, id, key, name, and title are returned by default if this parameter isn’t specified

Tags to return additional data. To return all additional data, use the tag ‘all’.

Possible values:

all

allytips

altimages

blurb

enemytips

image

info

lore

partype

passive

recommended

skins

spells

stats

tags

Sample output:

{:ok,
 %{"id" => 115, "key" => "Ziggs", "name" => "Ziggs",
   "title" => "the Hexplosives Expert"}}

champions(region, opts \\ %{})

Contains champion data for all champions.

Default opts:

locale: default varies based on region

Locale code for returned data (e.g., en_US, es_ES).

version: latest version based on region

Data dragon version for returned data.

dataById: false

If specified as true, the returned data map will use the champions’ IDs as the keys. If not specified or specified as false, the returned data map will use the champions’ keys instead.

champData: Only type, version, data, id, key, name, and title are returned by default if this parameter isn’t specified

Tags to return additional data. To return all additional data, use the tag ‘all’.

Possible values:

all

allytips

altimages

blurb

enemytips

image

info

lore

partype

passive

recommended

skins

spells

stats

tags

Sample output:

{:ok,
 %{"data" => %{"Ziggs" => %{"id" => 115, "key" => "Ziggs",
       "name" => "Ziggs", "title" => "the Hexplosives Expert"},
  ...
  },
 "type" => "champion", "version" => "7.4.3"}}

item(region, item_id, opts \\ %{})

Contains item data; retrieved by item id.

Default opts:

locale: default varies based on region

Locale code for returned data (e.g., en_US, es_ES).

version: latest version based on region

Data dragon version for returned data.

itemData: Only type, version, basic, data, id, name, plaintext, group, and description are returned by default if this parameter isn’t specified.

Tags to return additional data. To return all additional data, use the tag ‘all’.

Possible values:

all

colloq

consumeOnFull

consumed

depth

effect

from

gold

groups

hideFromAll

image

inStore

into

maps

requiredChampion

sanitizedDescription

specialRecipe

stacks

stats

tags

tree

Sample output:

{:ok,
 %{"description" => "<stats>+500 Health<br>+60 Armor<br>-10% Damage taken from Critical Strikes</stats><br><br><unique>UNIQUE Passive - Cold Steel:</unique> When hit by basic attacks, reduces the attacker's Attack Speed by 15% (1 second duration).<br><active>UNIQUE Active:</active> Slows the Movement Speed of nearby enemy units by 35% for 4 seconds (60 second cooldown).",
   "id" => 3143, "name" => "Randuin's Omen",
   "plaintext" => "Greatly increases defenses, activate to slow nearby enemies"}}

items(region, opts \\ %{})

Contains item data for all items.

Default opts:

locale: default varies based on region

Locale code for returned data (e.g., en_US, es_ES).

version: latest version based on region

Data dragon version for returned data.

itemListData: Only type, version, basic, data, id, name, plaintext, group, and description are returned by default if this parameter isn’t specified.

Tags to return additional data. To return all additional data, use the tag ‘all’.

Possible values:

all

colloq

consumeOnFull

consumed

depth

effect

from

gold

groups

hideFromAll

image

inStore

into

maps

requiredChampion

sanitizedDescription

specialRecipe

stacks

stats

tags

tree

Sample output:

{:ok,
 %{"basic" => %{"colloq" => "", "consumeOnFull" => false,
     "consumed" => false, "depth" => 1, "description" => "",
     "from" => [],
     "gold" => %{"base" => 0, "purchasable" => false, "sell" => 0,
       "total" => 0}, "group" => "", "hideFromAll" => false, "id" => 0,
     "image" => nil, "inStore" => true, "into" => [],
     "maps" => %{"1" => true, "10" => true, "12" => true, "8" => true},
     "name" => "", "plaintext" => "", "requiredChampion" => "",
     "rune" => %{"isRune" => false, "tier" => "1", "type" => "red"},
     "sanitizedDescription" => "", "specialRecipe" => 0, "stacks" => 1,
     "stats" => %{"FlatArmorMod" => 0.0, "PercentBlockMod" => 0.0,
       "rFlatMagicDamageModPerLevel" => 0.0, "PercentArmorMod" => 0.0,
       "rFlatSpellBlockModPerLevel" => 0.0,
       "rFlatEnergyModPerLevel" => 0.0, "FlatMPRegenMod" => 0.0,
       "rFlatArmorModPerLevel" => 0.0, "FlatAttackSpeedMod" => 0.0,
       "FlatBlockMod" => 0.0, "FlatHPRegenMod" => 0.0,
       "rFlatGoldPer10Mod" => 0.0, "rPercentTimeDeadMod" => 0.0,
       "PercentHPRegenMod" => 0.0,
       "rFlatMovementSpeedModPerLevel" => 0.0,
       "rFlatHPModPerLevel" => 0.0, "FlatEnergyRegenMod" => 0.0,
       "rFlatMPModPerLevel" => 0.0, "PercentSpellVampMod" => 0.0,
       "rFlatArmorPenetrationMod" => 0.0,
       "rPercentMovementSpeedModPerLevel" => 0.0, "rFlatDodgeMod" => 0.0, 
       "rFlatMPRegenModPerLevel" => 0.0,
       "rPercentMagicPenetrationModPerLevel" => 0.0,
       "PercentDodgeMod" => 0.0, "FlatCritChanceMod" => 0.0, ...},
     "tags" => []},
   "data" => %{"3124" => %{"description" => "<stats>+35 Attack Damage<br>+50 Ability Power<br>+25% Attack Speed</stats><br><br><passive>Passive: </passive>Basic attacks deal an additional 15 magic damage on hit.<br><unique>UNIQUE Passive:</unique> Basic attacks grant +8% Attack Speed, +3 Attack Damage, and +4 Ability Power for 5 seconds (stacks up to 6 times). While you have 6 stacks, gain <unlockedPassive>Guinsoo's Rage</unlockedPassive>.<br><br><unlockedPassive>Guinsoo's Rage:</unlockedPassive> Every other basic attack will trigger on hit effects an additional time.",
       "id" => 3124, "name" => "Guinsoo's Rageblade",
       "plaintext" => "Increases Ability Power and Attack Damage"},
     "2050" => %{"description" => "<consumable>Click to Consume:</consumable> Places an invisible ward that reveals the surrounding area for 60 seconds.",
       "id" => 2050, "name" => "Explorer's Ward"},
   ...
   },
 "type" => "item",
 "version" => "7.4.3"}}

language_strings(region, opts \\ %{})

Contains .

Sample output:

languages(region)

Contains .

Sample output:

map(region, opts \\ %{})

Contains .

Sample output:

masteries(region, opts \\ %{})

Contains .

Sample output:

mastery(region, id, opts \\ %{})

Contains .

Sample output:

realm(region)

Contains .

Sample output:

rune(region, id, opts \\ %{})

Contains .

Sample output:

runes(region, opts \\ %{})

Contains .

Sample output:

summoner_spell(region, id, opts \\ %{})

Contains .

Sample output:

summoner_spells(region, opts \\ %{})

Contains .

Sample output:

versions(region)

Contains .

Sample output: