Chemist v0.1.0 Chemist.Summoner

Retrieve summoner data from Riot API and transform it into an Elixir friendly format

Functions

id_by_name(region, name)

masteries_by_id(region, id)

name_by_id(region, id)

runes_by_id(region, id)

summoner_by_id(region, id)

summoner_by_name(region, summoner)

Return a tuple with the values { :ok, data } where data is a map of returned summoner attributes.

Uses “summoner-v1.4 API.”